requirements.txt
pillow 
matplotlib 
numpy
torch torchvision --index-url https://download.pytorch.org/whl/cpu